requirements.txt
lib==4.0.0
numpy==1.23.5
open3d==0.17.0
Pillow==9.4.0
Pillow==9.5.0
scipy==1.10.1
torch==2.0.0
